+++ graph/adapter-output-in-memory/src/main/kotlin/eu/tib/orkg/prototype/statements/adapter/output/inmemory/Extensions.kt
@@ -4,11 +4,23 @@ import eu.tib.orkg.prototype.statements.domain.model.ExactSearchString
 import eu.tib.orkg.prototype.statements.domain.model.FuzzySearchString
 import eu.tib.orkg.prototype.statements.domain.model.SearchString
 
-internal fun String.matches(searchString: SearchString): Boolean =
-    when (searchString) {
-    is ExactSearchString -> equals(searchString.value, ignoreCase = true)
+internal fun String.matches(searchString: SearchString): Boolean = when (searchString) {
+    is ExactSearchString -> equals(searchString.query, ignoreCase = true)
     is FuzzySearchString -> {
-        val unescaped = searchString.value.replace(Regex("""\\([~\[\]{}:^])"""), "$1")
-        contains(unescaped, ignoreCase = true)
+        val searchWords = searchString.query
+            .split(" AND ")
+            .map { it.replace(Regex("\\*$"), "") }
+        val words = split(" ")
+        searchWords.all { searchWord ->
+            when {
+                searchWord.startsWith("-") -> words.all {
+                    !it.startsWith(searchWord.substring(1))
+                }
+                searchWord.startsWith("+") -> words.any {
+                    it.startsWith(searchWord.substring(1))
+                }
+                else -> words.any { it.startsWith(searchWord) }
+            }
+        }
     }
 }