+++ rest-api-server/src/test/kotlin/eu/tib/orkg/prototype/statements/domain/model/SearchStringTest.kt
@@ -1,26 +1,61 @@
 package eu.tib.orkg.prototype.statements.domain.model
 
 import io.kotest.matchers.shouldBe
+import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 
 internal class SearchStringTest {
-    @Test
-    fun `Special characters escaped correctly when not escaped`() {
-        " ~ [ ] { } ^ : ".escapeFuzzySearchString() shouldBe """ \~ \[ \] \{ \} \^ \: """
-        "a~b[c]d{e}f^g:h".escapeFuzzySearchString() shouldBe """a\~b\[c\]d\{e\}f\^g\:h"""
-        "~[]{}^:".escapeFuzzySearchString() shouldBe """\~\[\]\{\}\^\:"""
+    @Nested
+    inner class Fuzzy {
+        @Nested
+        inner class Query {
+            @Test
+            fun `Escapes special characters`() {
+                FuzzySearchString("""+-&|!(){}[]^"~*?:\""").query shouldBe """\+\-\&\|\!\(\)\{\}\[\]\^\"\~\*\?\:\\"""
+            }
+
+            @Test
+            fun `Allows + and - when used as a word prefix`() {
+                FuzzySearchString("""+abc.""").query shouldBe """+abc."""
+                FuzzySearchString("""-def.""").query shouldBe """-def."""
+            }
+
+            @Test
+            fun `Inserts AND operator between words`() {
+                FuzzySearchString("""some. label.""").query shouldBe """some. AND label."""
+            }
+
+            @Test
+            fun `Inserts wildcards after words`() {
+                FuzzySearchString("""some label""").query shouldBe """some* AND label*"""
+            }
+        }
+
+        @Nested
+        inner class Input {
+            @Test
+            fun `Normalizes the input`() {
+                FuzzySearchString("""  some  label ${'\t'} """).input shouldBe """some label"""
+            }
+        }
     }
 
-    @Test
-    fun `Already escaped special characters are not escaped again`() {
-        """ \~ \[ \] \{ \} \^ \: """.escapeFuzzySearchString() shouldBe """ \~ \[ \] \{ \} \^ \: """
-        """a\~b\[c\]d\{e\}f\^g\:h""".escapeFuzzySearchString() shouldBe """a\~b\[c\]d\{e\}f\^g\:h"""
-        """\~\[\]\{\}\^\:""".escapeFuzzySearchString() shouldBe """\~\[\]\{\}\^\:"""
-    }
+    @Nested
+    inner class Exact {
+        @Nested
+        inner class Query {
+            @Test
+            fun `Escapes special characters`() {
+                ExactSearchString("""+-&|!(){}[]^"~*?:\""").query shouldBe """\+\-\&\|\!\(\)\{\}\[\]\^\"\~\*\?\:\\"""
+            }
+        }
 
-    @Test
-    fun `Leading special characters are escaped but are not escaped inside the string`() {
-        """???""".escapeFuzzySearchString() shouldBe """\???"""
-        """***""".escapeFuzzySearchString() shouldBe """\***"""
+        @Nested
+        inner class Input {
+            @Test
+            fun `Normalizes the input`() {
+                FuzzySearchString("""  some  label ${'\t'} """).input shouldBe """some label"""
+            }
+        }
     }
 }